BUS-9965 feat: Refatoração de código para substituição da estrutura de dados Map pelo banco H2(Work in Progress part. 6)
