+++ src/main/java/com/pagsestagio/movieapi/controller/requisicao/FilmeRequisicao.java
@@ -1,0 +1,12 @@
+package com.pagsestagio.movieapi.controller.requisicao;
+
+import com.pagsestagio.movieapi.model.Filme;
+
+public record FilmeRequisicao(Integer identificador, String nomeFilme) {
+
+    public FilmeRequisicao(Filme filme){
+        this(filme.getId(), filme.getNomeFilme());
+
+    }
+
+}
