+++ src/main/java/com/pagsestagio/movieapi/controller/resposta/FilmeResposta.java
@@ -1,0 +1,4 @@
+package com.pagsestagio.movieapi.controller.resposta;
+
+public interface FilmeResposta {
+}
