+++ src/main/java/com/pagsestagio/movieapi/controller/resposta/FilmeRespostaRetornaFilmeOuExcecaoV1.java
@@ -1,0 +1,7 @@
+package com.pagsestagio.movieapi.controller.resposta;
+
+
+@Deprecated
+public record FilmeRespostaRetornaFilmeOuExcecaoV1(Integer idLegado, String nomeFilme, String mensagemStatus) implements FilmeResposta {
+
+}
