+++ src/main/java/com/pagsestagio/movieapi/controller/resposta/FilmeRespostaRetornaFilmeOuMensagem.java
@@ -1,0 +1,6 @@
+package com.pagsestagio.movieapi.controller.resposta;
+
+
+import java.util.UUID;
+
+public record FilmeRespostaRetornaFilmeOuMensagem(UUID idpublico, String nomeFilme, String mensagemStatus) implements FilmeResposta {}
