--- src/main/java/com/pagsestagio/movieapi/latestversion/controller/resposta/FilmeResposta.java
@@ -1,4 +1,0 @@
-package com.pagsestagio.movieapi.latestversion.controller.resposta;
-
-public interface FilmeResposta {
-}
