--- src/main/java/com/pagsestagio/movieapi/latestversion/controller/resposta/FilmeRespostaRetornaFilmeOuMensagem.java
@@ -1,6 +1,0 @@
-package com.pagsestagio.movieapi.latestversion.controller.resposta;
-
-
-import java.util.UUID;
-
-public record FilmeRespostaRetornaFilmeOuMensagem(UUID idpublico, String nomeFilme, String mensagemStatus) implements FilmeResposta {}
