--- src/main/java/com/pagsestagio/movieapi/latestversion/model/FilmeDTO.java
@@ -1,6 +1,0 @@
-package com.pagsestagio.movieapi.latestversion.model;
-
-import java.util.UUID;
-
-public record FilmeDTO(UUID idPublico, String nomeFilme) {
-}
