--- src/main/java/com/pagsestagio/movieapi/latestversion/model/resultado/FilmeResultado.java
@@ -1,4 +1,0 @@
-package com.pagsestagio.movieapi.latestversion.model.resultado;
-
-public interface FilmeResultado {
-}
