--- src/main/java/com/pagsestagio/movieapi/latestversion/model/resultado/FilmeResultadoRetornaFilmeOuMensagem.java
@@ -1,7 +1,0 @@
-package com.pagsestagio.movieapi.latestversion.model.resultado;
-
-
-import java.util.UUID;
-
-public record FilmeResultadoRetornaFilmeOuMensagem(UUID idpublico, String nomeFilme, String mensagemStatus) implements FilmeResultado {
-}
