+++ src/main/java/com/pagsestagio/movieapi/model/FilmeDTO.java
@@ -1,0 +1,6 @@
+package com.pagsestagio.movieapi.model;
+
+import java.util.UUID;
+
+public record FilmeDTO(Integer id, Integer idLegado, UUID idPublico, String nomeFilme) {
+}
