+++ src/main/java/com/pagsestagio/movieapi/model/resultado/FilmeResultado.java
@@ -1,0 +1,4 @@
+package com.pagsestagio.movieapi.model.resultado;
+
+public interface FilmeResultado {
+}
