+++ src/main/java/com/pagsestagio/movieapi/model/resultado/FilmeResultadoRetornaFilmeOuExcecaoV1.java
@@ -1,0 +1,5 @@
+package com.pagsestagio.movieapi.model.resultado;
+
+@Deprecated
+public record FilmeResultadoRetornaFilmeOuExcecaoV1(Integer id, String nomeFilme, String mensagemStatus) implements FilmeResultado{
+}
