+++ src/main/java/com/pagsestagio/movieapi/model/resultado/FilmeResultadoRetornaFilmeOuMensagem.java
@@ -1,0 +1,7 @@
+package com.pagsestagio.movieapi.model.resultado;
+
+
+import java.util.UUID;
+
+public record FilmeResultadoRetornaFilmeOuMensagem(UUID idpublico, String nomeFilme, String mensagemStatus) implements FilmeResultado {
+}
