+++ src/main/java/com/pagsestagio/movieapi/model/resultado/FilmeResultadoRetornaListaDeFilmesOuExcecaoV1.java
@@ -1,0 +1,8 @@
+package com.pagsestagio.movieapi.model.resultado;
+
+import com.pagsestagio.movieapi.model.Filme;
+
+import java.util.List;
+@Deprecated
+public record FilmeResultadoRetornaListaDeFilmesOuExcecaoV1(List<Filme> listaDeFilmes, String mensagemStatus) implements FilmeResultado{
+}
