--- src/main/java/com/pagsestagio/movieapi/oldversion/controller/resposta/FilmeRespostaOldVersion.java
@@ -1,5 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.controller.resposta;
-
-@Deprecated
-public interface FilmeRespostaOldVersion {
-}
