--- src/main/java/com/pagsestagio/movieapi/oldversion/controller/resposta/FilmeRespostaRetornaFilmeOuExcecaoOldVersion.java
@@ -1,6 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.controller.resposta;
-
-@Deprecated
-public record FilmeRespostaRetornaFilmeOuExcecaoOldVersion(Integer id, String nomeFilme, String mensagemStatus) implements FilmeRespostaOldVersion {
-
-}
