--- src/main/java/com/pagsestagio/movieapi/oldversion/model/FilmeDTOOldVersion.java
@@ -1,5 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.model;
-
-@Deprecated
-public record FilmeDTOOldVersion(Integer identificador, String nomeFilme) {
-}
