--- src/main/java/com/pagsestagio/movieapi/oldversion/model/FilmeOldVersion.java
@@ -1,5 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.model;
-
-@Deprecated
-public record FilmeOldVersion(Integer identificador, String nomeFilme){
-}
