--- src/main/java/com/pagsestagio/movieapi/oldversion/model/resultado/FilmeResultadoOldVersion.java
@@ -1,5 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.model.resultado;
-
-@Deprecated
-public interface FilmeResultadoOldVersion {
-}
