--- src/main/java/com/pagsestagio/movieapi/oldversion/model/resultado/FilmeResultadoRetornaFilmeOuExcecaoOldVersion.java
@@ -1,5 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.model.resultado;
-
-@Deprecated
-public record FilmeResultadoRetornaFilmeOuExcecaoOldVersion(Integer id, String nomeFilme, String mensagemStatus) implements FilmeResultadoOldVersion {
-}
