--- src/main/java/com/pagsestagio/movieapi/oldversion/model/resultado/FilmeResultadoRetornaListaDeFilmesOuExcecaoOldVersion.java
@@ -1,9 +1,0 @@
-package com.pagsestagio.movieapi.oldversion.model.resultado;
-
-import com.pagsestagio.movieapi.oldversion.model.FilmeDTOOldVersion;
-
-import java.util.List;
-
-@Deprecated
-public record FilmeResultadoRetornaListaDeFilmesOuExcecaoOldVersion(List<FilmeDTOOldVersion> listaDeFilmes, String mensagemStatus) implements FilmeResultadoOldVersion {
-}
